1.17.13 #1030 laser render globally. also refactor inner class
