+++ src/main/java/com/lothrazar/cyclicmagic/block/beaconempty/TileEntityBeaconRedstoneRenderer.java
@@ -38,18 +38,19 @@ public class TileEntityBeaconRedstoneRenderer extends TileEntitySpecialRenderer<
 
   @Override
   public void render(TileEntityBeaconPowered te, double x, double y, double z, float partialTicks, int destroyStage, float alpha) {
+
     this.renderBeacon(x, y, z, partialTicks, te.shouldBeamRender(), te.getBeamSegments(),
         te.getWorld().getTotalWorldTime());
   }
 
-  public void renderBeacon(double x, double y, double z, double partialTicks, double textureScale, List<TileEntityBeaconPowered.BeamSegment> beamSegments, double totalWorldTime) {
+  public void renderBeacon(double x, double y, double z, double partialTicks, double textureScale, List<BeamSegment> beamSegments, double totalWorldTime) {
     GlStateManager.alphaFunc(516, 0.1F);
     this.bindTexture(TEXTURE_BEACON_BEAM);
     if (textureScale > 0.0D) {
       GlStateManager.disableFog();
       int i = 0;
       for (int j = 0; j < beamSegments.size(); ++j) {
-        TileEntityBeaconPowered.BeamSegment tileentitybeacon$beamsegment = beamSegments.get(j);
+        BeamSegment tileentitybeacon$beamsegment = beamSegments.get(j);
         TileEntityBeaconRenderer.renderBeamSegment(x, y, z, partialTicks, textureScale, totalWorldTime, i, tileentitybeacon$beamsegment.getHeight(), tileentitybeacon$beamsegment.getColors());
         i += tileentitybeacon$beamsegment.getHeight();
       }
@@ -59,6 +60,6 @@ public class TileEntityBeaconRedstoneRenderer extends TileEntitySpecialRenderer<
 
   @Override
   public boolean isGlobalRenderer(TileEntityBeaconPowered te) {
-    return true;
+    return true;//should link to tile entity setSetRenderGlobally
   }
 }
