+++ update.json
@@ -304,6 +304,6 @@
 	,"1.17.10":"Randomizer can no longer move unbreakable blocks (such as bedrock).  Catch and handle errors from 3rd party recipes when autocrafter is scanning recipe list, to prevent game crashes.  Added a config control to disable GPS Marker.  Updated zh_CN.lang by mcBegins2Snow.  Fixed the sprinkler sometimes grown crops client-side-only by fuami.  Nether and End ores now have extra ore dictionary entries added on top of usual, for example both 'oreCobalt' and new 'oreEndCobalt', added by josephcsible .  The /nbtprint command now also prints any ore dict entries for the held item.  Quickdraw enchant now works in offhand.  Multi-shot enchant: rebalanced damage of extra arrows with main based on charge.  Dehydrator fixed some recipes not processing such as saplings.  Dehydrator now stops consuming power if it has no recipe to work on.  Harvesting three tall crops now breaks the top block safely (Simple Corn).  Heart containers checking against config limit improved for compatibility.  Added Swim Speed II potion effect.  Machines and cables now compute energy serverside only, which is an optimization and fixes issues with some cables (such as Leadstone circuits).  Anvils no longer deal damage when you run into them at head height and all have the anvil sound.  "
 	,"1.17.11":"Updated some JEI information.  Updated language files.  Fixed issue with a keybinding with fastcraft.  Harvester compatibility Red Orchid.  Responsive scaffolding drops correct number of blocks.  When you shift-click a GPS Marker in a Transfer Node it now goes in the correct 1slot.  Ender Book has a new return feature, you can go back to your previous position that was saved on the previous teleport.  Exchange scepters now have two additional settings: 9x1 and 1x9 for column building.  New config optional for SFX block, if changed the 'sound_player_everybody' flag will instead make only the person using the block able to hear the sound, option to prevent sound-griefing.  "
 	,"1.17.12":"Fixed startup crash on servers with fastcraft.  Fix a crash & handle errors in Item extraction cable when using some types of tile entities (Thermal Dynamics Vacuum pipe with resonant servo, possibly others).  Fixed Packager not accepting zenscript (crafttweaker) recipes using nbt tag data.   "
-	,"1.17.13":"Brought back a legacy feature where Charms can be repaired in an anvil with an item from the recipe (PR from trunksbomb).  Ender Book teleport refactored off the main thread (helps with chunk loading issues on servers).  "
+	,"1.17.13":"Brought back a legacy feature where Charms can be repaired in an anvil with an item from the recipe (PR from trunksbomb).  Ender Book teleport refactored off the main thread (helps with chunk loading issues on servers).  Laser Cell now continues to render the laser even if the block is out of the players vision (same logic as Beacon).  "
   }
 }
